+++ Flow/Changes.md
@@ -1,4 +1,15 @@
 # Utopia Flow - List of Changes
+## v1.8 (beta)
+### Breaking Changes
+- Collection Extensions: 
+    - toMultiMap(f1, f2) and toMultiMap() were replaced with toMultiMap(f1)(f2), toMultiMap(...) and asMultiMap
+        - Use of asMultiMap is discouraged since it currently only supports tuple vectors
+        - New toMultiMap methods support all IterableOnce instances
+        
+### Fixes
+- **tryMap**(...) in CollectionExtensions now uses buildFrom and not Factory. This should result in better result 
+collection types when used.
+
 ## v1.7
 
 ### Major Changes